# openapi/apexrouter-v1.yaml
openapi: 3.1.0
info:
  title: ApexRouter-RS
  version: 0.1.0
  summary: One local router in front of every OpenAI-compatible endpoint you can reach.
  description: >-
    The machine-readable form of ARCHITECTURE.md section 6. Two listeners: the proxy plane
    (the OpenAI/Anthropic drop-in surface) and the control plane (rig, fit, routes, endpoints,
    providers, vast, hf, jobs). Every operation carries an `x-listener` extension naming which
    listener serves it; `/health` is the only path served by both. Prose lives in docs/API.md;
    routing semantics live in docs/ROUTING.md. This file is diffed against the axum route table
    by crates/apexrouter-server/tests/openapi_routes.rs on every CI run.
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  contact:
    name: ApexRouter-RS
    url: https://github.com/buckster123/ApexRouter-RS
servers:
  - url: http://127.0.0.1:8888
    description: 'proxy listener. `http://127.0.0.1:8888/v1` is equally valid: a doubled leading /v1 is collapsed.'
  - url: http://127.0.0.1:2739
    description: control listener.
...
security:
  - bearerAuth: []
  - tokenHeader: []
  - tokenQuery: []
paths:
...
  /v1/providers/{id}:
    get:
      operationId: getProvider
      x-listener: [control]
      tags: [providers]
      summary: One provider's status.
      parameters: [{ $ref: '#/components/parameters/ProviderId' }]
      responses:
        '200':
          description: The status row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderStatus' }
        '404': { $ref: '#/components/responses/ControlError' }
    put:
      operationId: putProvider
      x-listener: [control]
      tags: [providers]
      summary: Set base_url and exactly one credential form.
      description: >-
        A key the user typed goes to credentials.toml at 0600, never to config.toml. api_key_env and
        api_key_file are references; nothing is copied.
      parameters: [{ $ref: '#/components/parameters/ProviderId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderPut' }
      responses:
        '200':
          description: The updated status row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderStatus' }
        '400': { $ref: '#/components/responses/ControlError' }
  /v1/providers/{id}/test:
    post:
      operationId: testProvider
      x-listener: [control]
      tags: [providers]
      summary: Connection probe, plus an optional 16-token completion.
      description: >-
        `model` defaults to the first model the connection probe listed, never the hardcoded "x"
        that 400s on a managed provider.
      parameters:
        - $ref: '#/components/parameters/ProviderId'
        - name: completion
          in: query
          required: false
          description: Also send a 16-token completion, not just a connection probe.
          schema: { type: integer, format: int64 }
        - name: model
          in: query
          required: false
          description: Which model the completion should name.
          schema: { type: string }
      responses:
        '200':
          description: One row per probe.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CheckResult' }
  /v1/providers/{id}/models:
    get:
      operationId: providerModels
      x-listener: [control]
      tags: [providers]
      summary: The live catalogue, grouped by org.
      parameters: [{ $ref: '#/components/parameters/ProviderId' }]
      responses:
        '200':
          description: Upstream models.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UpstreamModel' }
        '503': { $ref: '#/components/responses/ControlError' }

  # =====================================================================================
  # control listener: vast
  # =====================================================================================
...
  /v1/compare:
    post:
      operationId: compare
      x-listener: [control]
      tags: [observability]
      summary: The same prompt against N aliases, at the same time.
      description: Blocking by default, because the answer IS the comparison.
      parameters: [{ $ref: '#/components/parameters/NoWait' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompareRequest' }
      responses:
        '200':
          description: One row per alias.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CompareRow' }
        '202':
          description: Accepted; the job's result is the same array.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRecord' }
...
  /v1/migrate:
    post:
      operationId: migrate
      x-listener: [control]
      tags: [observability]
      summary: Import ~/.vastai-gguf. dry_run writes nothing.
      description: >-
        --apply imports providers as credential REFERENCES; the real Together key is not copied.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MigrateRequest' }
      responses:
        '200':
          description: A MigrationPlan under dry_run, a MigrationReport otherwise.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MigrationPlan'
                  - $ref: '#/components/schemas/MigrationReport'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Authorization: Bearer <token>. Scopes read|write|admin are derived from (path, method).'
    tokenHeader:
      type: apiKey
      in: header
      name: X-ApexRouter-Token
      description: The same token, as a header.
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: The same token, as a query parameter. Never recorded in a tracing span.
  parameters:
...
    NoWait:
      name: no_wait
      in: query
      required: false
      description: Return a JobRecord immediately instead of waiting.
      schema: { type: boolean }
...
    ProviderId:
      name: id
      in: path
      required: true
      description: 'together, vast-gguf, hf, …'
      schema: { type: string }
...
  responses:
...
    ControlError:
      description: A control-plane ErrorEnvelope. Clients branch on error.kind, never on prose.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    BackendId:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      description: A validated slug. An empty or malformed id is not constructible.
    Alias:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
      description: The string a client puts in "model".
...
    TokenCount:
      type: object
      required: [kind, n]
      properties:
        kind: { type: string, enum: [reported, estimated] }
        n: { type: integer, format: int64 }

    # ---- errors -----------------------------------------------------------------------
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    ErrorBody:
      type: object
      required: [kind, message]
      properties:
        kind:
          type: string
          description: 'The discriminator, so no exit codes had to be invented.'
          examples: [not_found, invalid_request, conflict, port_in_use, unauthorized, forbidden,
                     upstream_unavailable, approval_required, internal]
        message: { type: string }
        param: { type: [string, 'null'] }
        code: { type: [string, 'null'] }
...
    JobRecord:
      type: object
      required: [id, kind, state, started_unix]
      properties:
        id: { type: string }
        kind: { type: string, examples: [hf.download, vast.rent, endpoint.start, compare, swap] }
        state:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
          description: Never Pending forever; even a JoinError from a panic flips it to failed.
        pct: { type: [number, 'null'] }
        message: { type: [string, 'null'] }
        started_unix: { type: integer, format: int64 }
        finished_unix: { type: [integer, 'null'], format: int64 }
        result: { description: 'The protocol type the operation produced, as JSON.' }
        error: { type: [string, 'null'] }

    # ---- control plane: rig and fit -----------------------------------------------------
...
    CredentialSource:
      type: object
      description: A DESCRIPTION of where a credential lives. Never key material.
      required: [kind]
      properties:
        kind: { type: string, enum: [none, env, file, managed, instance] }
        var: { type: string }
        path: { type: string }
        store: { type: string }
...
    UpstreamModel:
      type: object
      properties:
        id: { type: string }
        ctx: { type: [integer, 'null'], format: int64 }
        vision: { type: boolean }
...
    ProviderStatus:
      type: object
      required: [id, base_url, credential, credential_present]
      properties:
        id: { type: string }
        base_url:
          type: string
          description: 'A legacy api.together.xyz is used as-is and never rewritten to .ai.'
        credential: { $ref: '#/components/schemas/CredentialSource' }
        credential_present: { type: boolean, description: Whether resolving that source actually produced something. }
        models_cached: { type: integer, format: int64 }
        last_ok_unix: { type: [integer, 'null'], format: int64 }
        last_error: { type: [string, 'null'] }
        rate_limit: { type: [object, 'null'], additionalProperties: true }
    ProviderPut:
      type: object
      description: At most one of api_key, api_key_env, api_key_file.
      properties:
        base_url: { type: [string, 'null'] }
        api_key: { type: [string, 'null'], description: Typed by a human → credentials.toml at 0600. }
        api_key_env: { type: [string, 'null'], description: A reference. Nothing is copied. }
        api_key_file: { type: [string, 'null'], description: A reference. Nothing is copied. }
    CheckResult:
      type: object
      required: [id, label, status, ms, detail]
      properties:
        id: { type: string }
        label: { type: string }
        status: { type: string, enum: [pass, warn, fail, skipped] }
        ms: { type: integer, format: int64 }
        detail: { type: string }
        fix: { type: [string, 'null'], description: 'An actionable line, never prose.' }

    # ---- control plane: vast ---------------------------------------------------------------
...
    CompareRequest:
      type: object
      required: [prompt]
      properties:
        aliases: { type: array, items: { type: string } }
        prompt: { type: string }
        max_tokens: { type: [integer, 'null'], format: int64 }
    CompareRow:
      type: object
      properties:
        alias: { $ref: '#/components/schemas/Alias' }
        backend: { oneOf: [{ $ref: '#/components/schemas/BackendId' }, { type: 'null' }] }
        model: { type: string }
        ok: { type: boolean }
        ms: { type: integer, format: int64 }
        ttft_ms: { type: [integer, 'null'], format: int64 }
        tok_per_s: { type: [number, 'null'] }
        prompt_tokens:
          description: 'The REAL number from the response, never word_count × 1.3.'
          oneOf: [{ $ref: '#/components/schemas/TokenCount' }, { type: 'null' }]
        text: { type: [string, 'null'] }
      additionalProperties: true
...
    MigrateRequest:
      type: object
      required: [dry_run]
      properties:
        dry_run: { type: boolean, description: 'true returns a MigrationPlan and writes NOTHING.' }
    MigrationItem:
      type: object
      properties:
        what: { type: string, examples: [recipe, usage row, known_fork, provider, instance] }
        from: { type: string }
        action: { type: string, enum: [import, skip, warn] }
        detail: { type: string, description: 'The reason, per row.' }
    MigrationPlan:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/MigrationItem' } }
        source_paths: { type: array, items: { type: string } }
    MigrationReport:
      type: object
      properties:
        imported: { type: integer, format: int64 }
        skipped: { type: integer, format: int64 }
        warnings: { type: array, items: { type: string } }